feat: credit amount 조회 함수의 자료형을 Integer로 수정

+++ main-modules/credit/application/src/main/java/com/sidework/credit/application/port/in/CreditQueryUseCase.java
@@ -1,5 +1,5 @@
 package com.sidework.credit.application.port.in;
 
 public interface CreditQueryUseCase {
-    Long sumAmountByUser(Long userId);
+    Integer sumAmountByUser(Long userId);
 }
