+++ main-modules/credit/infrastructure/src/main/java/com/sidework/credit/persistence/adapter/CreditPersistenceAdapter.java
@@ -23,7 +23,7 @@ public class CreditPersistenceAdapter implements CreditOutPort {
     }
 
     @Override
-    public Long findAmountByUser(Long userId) {
+    public Integer findAmountByUser(Long userId) {
         return repo.findSumAmountByUserId(userId);
     }
